break: fix break display for lots of balls
For breaks more than 10 balls, switch to showing ball counts. This is a
bit like TV broadcast display of the balls potted.

Unfortunately a lot of fonts appear to have size variations in the
numbers, because 0 and 11-15 come from different unicode blocks. Anyway,
this is a start, may be improved later.